+++ CHANGELOG.md
@@ -6,10 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+<<<<<<< Updated upstream:CHANGELOG.md
 - Stuff in the works here
 - Change Keystroke to include LegalKey as validation
 - Move KeyLogger.simulate to KeySimulator
 - 
+=======
+- Implement LegalKey validation and equality
+- include a separate thread for monitoring the elapsed time in the Listener
+- Allow KeySimulator to run simulation while self.off == True where keystroke type events will not be executed
+- Implement proper logging instead of print statements
+- Implement functions to double check the equality of all Keystrokes in the KeyLogger and KeySimulator
+- Add complex logic in Keystroke to ensure non-ascii chars aren't typed ungodly fast
+>>>>>>> Stashed changes:unpack/CHANGELOG.md
 ## [1.1.1] - 2023-10-13
 
 ### Added